requirements.txt

Django==4.2.7
django-smtp-ssl==1.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
importlib-metadata 
importlib-resources==6.1.0
keras==2.14.0
keras-facenet==0.3.2
matplotlib==3.8.0
mediapipe==0.9.1.0
mpmath==1.3.0
mtcnn==0.1.1
numpy==1.26.1
onnx==1.14.1
onnxruntime==1.16.1
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.74
pandas==2.1.2
pickleshare 
Pillow==10.1.0
PyMySQL==1.1.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tf2onnx==1.15.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
yolov5==7.0.13

